+++ src-test/Transformation/TestTransformationAlgorithm-case-0.cpp
@@ -46,8 +46,8 @@ using namespace xobj;
  * Without root matrix
  */
 TEST(TestTransformationAlgorithm_case0, case1) {
-    const std::string path(std::string(TOTEXT(TestTransformationAlgorithm_case0)).append(".").append(TOTEXT(case1)).append(".obj"));
-    //-------------------
+    const auto fileName = XOBJ_PATH("TestTransformationAlgorithm_case0-case1.obj");
+    //-----------------------------
     // make out data and save to file
 
     ObjMain mainOut;
@@ -67,13 +67,15 @@ TEST(TestTransformationAlgorithm_case0, case1) {
     transformOut1.pMatrix.setPosition(Point3(-50.0f, 0.0f, 0.0f));
     transformOut2.pMatrix.setPosition(Point3(50.0f, 0.0f, 0.0f));
 
-    ASSERT_TRUE(mainOut.exportToFile(path));
+    ExportContext expContext(fileName);
+    ASSERT_TRUE(mainOut.exportObj(expContext));
 
     //-------------------
     // load data from file
 
     ObjMain mainIn;
-    ASSERT_TRUE(mainIn.importFromFile(path));
+    ImportContext impContext(fileName);
+    ASSERT_TRUE(mainIn.importObj(impContext));
 
     // extract data
     ObjLodGroup * lodIn = nullptr;
@@ -98,8 +100,8 @@ TEST(TestTransformationAlgorithm_case0, case1) {
  * With root matrix
  */
 TEST(TestTransformationAlgorithm_case0, case2) {
-    const std::string path(std::string(TOTEXT(TestTransformationAlgorithm_case0)).append(".").append(TOTEXT(case2)).append(".obj"));
-    //-------------------
+    const auto fileName = XOBJ_PATH("TestTransformationAlgorithm_case0-case2.obj");
+    //-----------------------------
     // make out data and save to file
 
     ObjMain mainOut;
@@ -116,13 +118,15 @@ TEST(TestTransformationAlgorithm_case0, case2) {
     transformOut1.pMatrix.setPosition(Point3(-50.0f, 0.0f, 0.0f));
     transformOut1.pMatrix.rotateDegreesY(+90.0f);
 
-    ASSERT_TRUE(mainOut.exportToFile(path));
+    ExportContext expContext(fileName);
+    ASSERT_TRUE(mainOut.exportObj(expContext));
 
     //-------------------
     // load data from file
 
     ObjMain mainIn;
-    ASSERT_TRUE(mainIn.importFromFile(path));
+    ImportContext impContext(fileName);
+    ASSERT_TRUE(mainIn.importObj(impContext));
 
     // extract data
     ObjLodGroup * lodIn = nullptr;