+++ src-test/Transformation/TestTransformationAlgorithm-case-5.cpp
@@ -50,8 +50,8 @@ using namespace xobj;
  * Animation moving the child along parent's X, so as the result in the world space the moving will be along Y axis.
  */
 TEST(TestTransformationAlgorithm_case5, case1) {
-    const std::string path(std::string(TOTEXT(TestTransformationAlgorithm_case5)).append(".").append(TOTEXT(case1)).append(".obj"));
-    //-------------------
+    const auto fileName = XOBJ_PATH("TestTransformationAlgorithm_case5-case1.obj");
+    //-----------------------------
     // make out data and save to file
 
     ObjMain mainOut;
@@ -74,13 +74,15 @@ TEST(TestTransformationAlgorithm_case5, case1) {
                                        AnimTransKey(-50.0f, 0.0f, 0.0f, -10.0f), AnimTransKey(50.0f, 0.0f, 0.0f, 10.0f), "trans1");
     TestUtils::createTestAnimTranslate(trOut2.pAnimTrans,
                                        AnimTransKey(-50.0f, 0.0f, 0.0f, -10.0f), AnimTransKey(50.0f, 0.0f, 0.0f, 10.0f), "trans2");
-    ASSERT_TRUE(mainOut.exportToFile(path));
+    ExportContext expContext(fileName);
+    ASSERT_TRUE(mainOut.exportObj(expContext));
 
     //-------------------
     // load data from file
 
     ObjMain mainIn;
-    ASSERT_TRUE(mainIn.importFromFile(path));
+    ImportContext impContext(fileName);
+    ASSERT_TRUE(mainIn.importObj(impContext));
 
     // extract data
     ObjLodGroup * lodIn = nullptr;
@@ -127,8 +129,8 @@ TEST(TestTransformationAlgorithm_case5, case1) {
  * Then rotate the whole scene back with global matrix.
  */
 TEST(TestTransformationAlgorithm_case5, case2) {
-    const std::string path(std::string(TOTEXT(TestTransformationAlgorithm_case5)).append(".").append(TOTEXT(case2)).append(".obj"));
-    //-------------------
+    const auto fileName = XOBJ_PATH("TestTransformationAlgorithm_case5-case2.obj");
+    //-----------------------------
     // make out data and save to file
 
     ObjMain mainOut;
@@ -152,13 +154,15 @@ TEST(TestTransformationAlgorithm_case5, case2) {
                                        AnimTransKey(-50.0f, 0.0f, 0.0f, -10.0f), AnimTransKey(50.0f, 0.0f, 0.0f, 10.0f), "trans1");
     TestUtils::createTestAnimTranslate(trOut2.pAnimTrans,
                                        AnimTransKey(-50.0f, 0.0f, 0.0f, -10.0f), AnimTransKey(50.0f, 0.0f, 0.0f, 10.0f), "trans2");
-    ASSERT_TRUE(mainOut.exportToFile(path));
+    ExportContext expContext(fileName);
+    ASSERT_TRUE(mainOut.exportObj(expContext));
 
     //-------------------
     // load data from file
 
     ObjMain mainIn;
-    ASSERT_TRUE(mainIn.importFromFile(path));
+    ImportContext impContext(fileName);
+    ASSERT_TRUE(mainIn.importObj(impContext));
 
     // extract data
     ObjLodGroup * lodIn = nullptr;