+++ src/objects/ImportContext.cpp
@@ -1,5 +1,5 @@
 /*
-**  Copyright(C) 2017, StepToSky
+**  Copyright(C) 2018, StepToSky
 **
 **  Redistribution and use in source and binary forms, with or without
 **  modification, are permitted provided that the following conditions are met:
@@ -27,57 +27,21 @@
 **  Contacts: www.steptosky.com
 */
 
-#include <gtest/gtest.h>
-#include "xpln/obj/ObjDrapedGroup.h"
-#include "TestUtils.h"
-#include "TestUtilsObjMesh.h"
-
-using namespace xobj;
+#include "xpln/obj/ImportContext.h"
 
 /**************************************************************************************************/
-////////////////////////////////////////////////////////////////////////////////////////////////////
+//////////////////////////////////////////* Functions */////////////////////////////////////////////
 /**************************************************************************************************/
 
-TEST(Draped, DISABLED_io) {
-    const auto fileName = XOBJ_PATH("Draped-io.obj");
-    //-------------------
-    // make out data and save to file
-    ObjMain mainOut;
-    TestUtils::setTestExportOptions(mainOut);
-    ObjLodGroup & lod1 = mainOut.addLod(new ObjLodGroup("l1", 0.0f, 100.0f));
-    ObjLodGroup & lod2 = mainOut.addLod(new ObjLodGroup("l2", 100.0f, 200.0f));
+void xobj::ImportContext::setInterruptor(IInterrupt * interruptor) {
+    mInterruptor.reset(interruptor);
+}
 
-    lod1.transform().addObject(TestUtilsObjMesh::createPyramidTestMesh("l1-m1"));
-    lod1.transform().addObject(TestUtilsObjMesh::createPyramidTestMesh("l1-m2"));
-
-    lod2.transform().addObject(TestUtilsObjMesh::createPyramidTestMesh("l2-m1"));
-    lod2.transform().addObject(TestUtilsObjMesh::createPyramidTestMesh("l2-m2"));
-
-    mainOut.pDraped.transform().addObject(TestUtilsObjMesh::createPyramidTestMesh("d1"));
-    mainOut.pDraped.transform().addObject(TestUtilsObjMesh::createPyramidTestMesh("d2"));
-
-    ExportContext expContext(fileName);
-    ASSERT_TRUE(mainOut.exportObj(expContext));
-
-    //-------------------
-    // load data from file
-
-    ObjMain mainIn;
-    ImportContext impContext(fileName);
-    ASSERT_TRUE(mainIn.importObj(impContext));
-
-    ObjLodGroup * lodIn1 = nullptr;
-    ObjLodGroup * lodIn2 = nullptr;
-    // One transform was optimized during export, it became LOD's transform
-    ASSERT_NO_FATAL_FAILURE(TestUtils::extractLod(mainIn, 0, lodIn1));
-    ASSERT_NO_FATAL_FAILURE(TestUtils::extractLod(mainIn, 1, lodIn2));
-
-    //-------------------
-    // check results
-
-    EXPECT_EQ(2, lodIn1->transform().objList().size());
-    EXPECT_EQ(2, lodIn2->transform().objList().size());
-    EXPECT_EQ(2, mainIn.pDraped.transform().objList().size());
+xobj::IInterrupt * xobj::ImportContext::interruptor() {
+    if (!mInterruptor) {
+        mInterruptor = std::make_unique<NoInterrupt>();
+    }
+    return mInterruptor.get();
 }
 
 /**************************************************************************************************/