+++ changelog.md
@@ -3,10 +3,15 @@ StepToSky: X-Plane Obj Library
 #### next
 
 - **Added** Classes for reading/writing datarefs and commands files.
+- **Added** Context for importing/exporting.
 - **Fixed** Fixed bug for attribute printing: 
             `AttrShiny`  
             `AttrBlend`  
             Default/disable values were not printed when those attributes were enabled and then disabled.
+
+##### Breaking backward compatibility:
+- **Changed:** Methods for export/import in ObjMain.
+- **Changed:** Under windows (MSVC) std::wstring is used as a path.
 
 ---------------------------------------------------------------------------
 #### 0.7.0-beta (10.10.2018)